+++ ui/src/pages/ServerError.vue
@@ -1,25 +1,22 @@
 <template>
-  <div class="bg-grey-2 q-px-xl column window-height items-center justify-around">
-    <div class="column full-width items-center justify-center q-gutter-md q-py-md row">
-      <q-img src="../assets/gear.svg" :width="gearSize" />
-      <div class="text-bold text-h1 text-primary">500</div>
-      <div class="column items-center">
-        <div class="text-bold text-h6 text-subtitle text-grey-9">Ops, ocorreu um erro interno no nosso servidor!</div>
-        <div class="text-bold text-center text-subtitle text-grey-9">Tente novamente em alguns instantes.</div>
-      </div>
-    </div>
-  </div>
+  <error-component v-bind="errorProps" />
 </template>
 
 <script>
-import screenMixin from '../mixins/screen'
+import ErrorComponent from './ErrorComponent'
 
 export default {
-  mixins: [screenMixin],
+  components: {
+    ErrorComponent
+  },
 
   computed: {
-    gearSize () {
-      return this.$_isSmall ? '90px' : '160px'
+    errorProps () {
+      return {
+        code: '500',
+        title: 'Ops, ocorreu um erro interno no nosso servidor!',
+        subtitle: 'Tente novamente em alguns instantes.'
+      }
     }
   }
 }