+++ app-extension/src/boot/error-pages.js
@@ -1,5 +1,6 @@
 import ForbiddenPage from '@bildvitta/quasar-ui-asteroid/src/pages/Forbidden.vue'
 import NotFound from '@bildvitta/quasar-ui-asteroid/src/pages/NotFound.vue'
+import ServerError from '@bildvitta/quasar-ui-asteroid/src/pages/ServerError.vue'
 
 export default function ({ router }) {
   const routes = [
@@ -7,6 +8,11 @@ export default function ({ router }) {
       name: 'Forbidden',
       path: '/',
       component: ForbiddenPage
+    },
+    {
+      name: 'ServerError',
+      path: '/',
+      component: ServerError
     }
   ]