+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Add
+- Added new prop `dialogProps` and `signaturePadProps` in `QasSignatureUploader`.
+- Added new HTML atribute `width` at `<canvas />` in `QasSignaturePad` and new props `width` and `height`.
+
 ## 2.17.1 - 2022-04-11
 
 ### Fixed