+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Add
-- Added new prop `dialogProps` and `signaturePadProps` in `QasSignatureUploader`.
-- Added new HTML atribute `width` at `<canvas />` in `QasSignaturePad` and new props `width` and `height`.
+- Added new prop `dialogProps` in `QasSignatureUploader`.
+- Added new HTML atribute `width` at `<canvas />` in `QasSignaturePad`.
+- Added computed values to set canvas size.
 
 ## 2.17.1 - 2022-04-11